feat: viewall permintaan pengiriman, view permintaan pengiriman, create permintaan pengiriman

+++ silogistik2106639485/src/main/java/apap/ti/silogistik2106639485/repository/PermintaanPengirimanDb.java
@@ -2,12 +2,10 @@ package apap.ti.silogistik2106639485.repository;
 
 import apap.ti.silogistik2106639485.model.PermintaanPengiriman;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PermintaanPengirimanDb extends JpaRepository<PermintaanPengiriman, BigInteger> {
+public interface PermintaanPengirimanDb extends JpaRepository<PermintaanPengiriman, Long> {
     
 }
